dinamicidade p aplicação
a página inicial agora se comunica com o bean jsf e recupera as
informações do banco de dados para inserir via js/jquery na página html.
os dados são transmitidos com json

+++ greenhouse/src/java/br/udesc/greenhouse/modelo/dao/core/NoticiaDAO.java
@@ -18,9 +18,9 @@ public interface NoticiaDAO {
 
     public void editar(Noticia a);
 
-    public void remover(int id);
+    public void remover(long id);
 
-    public Noticia pesquisar(int id);
+    public Noticia pesquisar(long id);
 
     public List<Noticia> listar();
